Clean up unnecessary code
