+++ Benchmark/BenchmarkCommon/BenchmarkMemoryManager.hpp
@@ -45,7 +45,7 @@ void BM_AllocFixedSize_stdnew(benchmark::State& state)
 void BM_AllocFixedSize_FixedSizeAllocator(benchmark::State& state)
 {
 	auto size = state.range(0);
-	SpaceGameEngine::MemoryManager::FixedSizeAllocator allocator(size, 4096, 4);
+	SpaceGameEngine::MemoryManager::LockedFixedSizeAllocator allocator(size, 4096, 4);
 	for (auto _ : state)
 	{
 		void* ptr = allocator.Allocate();
