+++ Source/Common/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(Common
         Private/Error.cpp
-        Private/MemoryManager.cpp
+        Private/Memory/LockedFixedSizeAllocator.cpp
         Private/SGEString.cpp
         Private/StringImplement/StdTStringImplement.cpp
         Private/Concurrent/Thread.cpp
